backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
tenacity>=8.2.3
pyyaml>=6.0.1
